# public/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Petshop API",
        "description": "API routes for the Petshop coding exercies",
        "contact": {
            "email": "chris.idakwo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:9900",
            "description": "Base API URL"
        }
    ],
    "paths": {
...
        "/api/v1/main/promotions": {
            "get": {
                "tags": [
                    "MainPage"
                ],
                "summary": "List all promotions",
                "description": "List all promotions",
                "operationId": "main/promotions/index",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sortBy",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "desc",
                        "in": "query",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok"
                    }
                }
            }
        },
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Login a user account",
                "description": "Login a user account",
                "operationId": "user/login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email address",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john.doe@email.com",
                                    "password": "password"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Logout a user account",
                "description": "Logout a user account",
                "operationId": "user/logout",
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/forgot-password": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Creates a token to reset a user password",
                "description": "Creates a token to reset a user password",
                "operationId": "user/forgot-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
        "/api/v1/user/reset-password-token": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Reset a user password with a token",
                "description": "Reset a user password with a token",
                "operationId": "user/reset-password",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "token",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "User reset token",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    }
                }
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT authentication token",
                "scheme": "bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}